Aggiunta CarrelloController

+++ Filiera_Francoletti_Belardinelli_Raiola/src/main/java/com/example/filiera_francoletti_belardinelli_raiola/model/Payment/Carrello.java
@@ -26,7 +26,7 @@ public class Carrello {
         }
     }
 
-    public void removeProduct(int id) {
+    public void removeProduct(Long id) {
         products.removeIf(product -> product.getId().intValue() == id);
     }
