+++ Filiera_Francoletti_Belardinelli_Raiola/src/main/java/com/example/filiera_francoletti_belardinelli_raiola/model/Users/Acquirente.java
@@ -40,7 +40,7 @@ public class Acquirente extends UtenteGenerico {
         return buyerHandler.pay(this).getInvoice();
     }
 
-    public void removeProduct(int id) {
+    public void removeProduct(Long id) {
         buyerHandler.removeProduct(id);
     }
