+++ Filiera_Francoletti_Belardinelli_Raiola/src/main/java/com/example/filiera_francoletti_belardinelli_raiola/service/HandlerAcquirente.java
@@ -33,7 +33,7 @@ public class HandlerAcquirente {
         return pagamento;
     }
 
-    public void removeProduct(int id) {
+    public void removeProduct(Long id) {
         this.shoppingCart.removeProduct(id);
     }
